Fix SpiceController::Connect sometimes returning success without connecting
SpiceController::Connect (the retry wrapper version) was using rc
uninitialized, and if rc happened to be 0 it would return success without
actually connecting.

Also zap the check for a negative retry count if someone is stupid enough
to do that we will never enter the loop and return the initial
(now initialized to -1) rc value.