build: deal with non-const return values of NPP_GetMIMEDescription()
Deal with const and non-const return value of NPP_GetMIMEDescription().

Although it is hard to find evidences of how this value is used, I
think the non-const value could return a static string. But I prefer
being on the safe side and return a strdup() to the risk of leaking
that string on older versions.

Fixes the build on RHEL6 with xulrunner-1.9.2.24-2.el6.