Switch from log4cpp to glib
GLib is used in the defacto library for desktop projects. It includes
logging facilities that can be filtered by domain and level, and can
be redirected or formated in various ways. Instead of pulling an extra
large depedency for a small plugin, let's use what is common to the
desktop.

Afaict, nobody else is using log4cpp in Fedora. The configuration file
is not user friendly. In comparison, you can just run
"G_MESSAGES_DEBUG=SpiceXPI firefox" and easily modify logging via
environment variable. There is no logging to file, and there is no
configuration file. If you want to redirect to file, you can do so via
redirection, or other mechanisms (ldpreload, environment..).