Get rid of nspr use
The only thing from nspr we use are typedef for bool, int32_t and
definitions of TRUE/FALSE. Any decent C/C++ compiler should provide
all of these. If this proves to be an issue, the NPAPI headers
we use to build the plugin have a nptypes.h header with the needed
defines for bool/true/false and int32_t.