Add glib Windows logging
As it's not very convenient to get logging output from the plugin
while firefox is running, this commit adds a dumb glib logging
implementation that writes the logging information to a file.
As this is not efficient at all, it's disabled by default unless
the SPICE_XPI_LOG_TO_FILE environment variable is set.