Make spice.xpi build optional
Now that we switched to using the headers from npapi-sdk to build
the spice-xpi plugin, the only reason we are depending on xulrunner
is to build an xpi package for the plugin for easier installation
in firefox. However, this .xpi is not installed by default nor
packaged by distributions, so it's not in wide use.

This commit makes its build optional as this has the side-effect
on making the xulrunner dependency optional as well, which is a
good thing as xulrunner is huge and not necessarily available
on all platforms we'd like to build spice-xpi on.